Fix errors in update-dependencies tool, always try to update both latest and specific Chisel versions (#5719)
