+++ eng/update-dependencies/ChiselUpdater.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using System.Net.Http;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Microsoft.DotNet.VersionTools.Dependencies;
@@ -20,20 +21,19 @@ internal static class ChiselUpdater
         Release rocksToolboxRelease = await GitHubHelper.GetLatestRelease("canonical", "rocks-toolbox");
 
         IEnumerable<IDependencyUpdater> chiselUpdaters = s_chiselArchitectures.SelectMany<string, IDependencyUpdater>(arch =>
-        {
-            string manifestVar = GetChiselManifestVariable("chisel", arch, "sha", dockerfileVersion);
-            return [
-                new ChiselReleaseShaUpdater(repoRoot, manifestVar, chiselRelease, arch),
-                new GitHubReleaseUrlUpdater(repoRoot, manifestVar, chiselRelease, DependencyInfoToUse, GetAssetRegex(arch))
-            ];
-        });
+            [
+                new ChiselReleaseShaUpdater(repoRoot, GetChiselManifestVariable("chisel", arch, "sha", "latest"), chiselRelease, arch),
+                new ChiselReleaseShaUpdater(repoRoot, GetChiselManifestVariable("chisel", arch, "sha", dockerfileVersion), chiselRelease, arch),
+                new GitHubReleaseUrlUpdater(repoRoot, GetChiselManifestVariable("chisel", arch, "url", "latest"), chiselRelease, DependencyInfoToUse, GetAssetRegex(arch)),
+                new GitHubReleaseUrlUpdater(repoRoot, GetChiselManifestVariable("chisel", arch, "url", dockerfileVersion), chiselRelease, DependencyInfoToUse, GetAssetRegex(arch))
+            ]);
 
         GitHubReleaseVersionUpdater rocksToolboxUpdater = new(repoRoot, "rocks-toolbox|latest|version", rocksToolboxRelease, DependencyInfoToUse);
 
         return [ ..chiselUpdaters, rocksToolboxUpdater ];
     }
 
-    private static string GetChiselManifestVariable(string product, string arch, string type, string dockerfileVersion = "latest")
+    public static string GetChiselManifestVariable(string product, string arch, string type, string dockerfileVersion = "latest")
     {
         return string.Join('|', [product, dockerfileVersion, ToManifestArch(arch), type]);
     }
@@ -41,4 +41,21 @@ internal static class ChiselUpdater
     private static Regex GetAssetRegex(string arch) => new(@"chisel_v\d+\.\d+\.\d+_linux_" + arch + @"\.tar\.gz");
 
     private static string ToManifestArch(string arch) => arch == "amd64" ? "x64" : arch;
+
+    private sealed class ChiselReleaseShaUpdater(string repoRoot, string variableName, Release release, string arch)
+        : GitHubReleaseUrlUpdater(repoRoot, variableName, release, DependencyInfoToUse, GetAssetRegex(arch))
+    {
+        private static readonly HttpClient s_httpClient = new();
+
+        protected override string? GetValue()
+        {
+            string? downloadUrl = base.GetValue();
+            if (downloadUrl is null)
+            {
+                return null;
+            }
+
+            return ChecksumHelper.ComputeChecksumShaAsync(s_httpClient, downloadUrl).Result;
+        }
+    }
 }
