+++ manifest.versions.json
@@ -84,10 +84,10 @@
     "chisel|9.0|arm|url": "$(chisel|latest|arm|url)",
     "chisel|9.0|arm64|url": "$(chisel|latest|arm64|url)",
     "chisel|8.0|x64|url": "$(chisel|latest|x64|url)",
-    "chisel|6.0|arm|url": "$(chisel|latest|arm|url)",
+    "chisel|8.0|arm|url": "$(chisel|latest|arm|url)",
     "chisel|8.0|arm64|url": "$(chisel|latest|arm64|url)",
     "chisel|6.0|x64|url": "$(chisel|latest|x64|url)",
-    "chisel|8.0|arm|url": "$(chisel|latest|arm|url)",
+    "chisel|6.0|arm|url": "$(chisel|latest|arm|url)",
     "chisel|6.0|arm64|url": "$(chisel|latest|arm64|url)",
 
     "chisel|latest|x64|sha": "5c863d41e9b0829a3af57de444bdb7d1638612e5caa0a289511b5f46ec5728795152381257d30bbff012bcb5c533b6cfbc17b17b93569fd5c0e194ce2c6e0fc5",
@@ -97,10 +97,10 @@
     "chisel|9.0|arm|sha": "$(chisel|latest|arm|sha)",
     "chisel|9.0|arm64|sha": "$(chisel|latest|arm64|sha)",
     "chisel|8.0|x64|sha": "$(chisel|latest|x64|sha)",
-    "chisel|6.0|arm|sha": "$(chisel|latest|arm|sha)",
-    "chisel|8.0|arm64|sha": "$(chisel|latest|arm64|sha)",
-    "chisel|6.0|x64|sha": "$(chisel|latest|x64|sha)",
     "chisel|8.0|arm|sha": "$(chisel|latest|arm|sha)",
+    "chisel|8.0|arm64|sha": "$(chisel|latest|arm64|sha)",
+    "chisel|6.0|arm|sha": "$(chisel|latest|arm|sha)",
+    "chisel|6.0|x64|sha": "$(chisel|latest|x64|sha)",
     "chisel|6.0|arm64|sha": "$(chisel|latest|arm64|sha)",
 
     "dotnet|6.0|product-version": "6.0.32",
